move formatPropertyValue to Models
