+++ lib/Map/TableDataSource.js
@@ -16,7 +16,7 @@ var destroyObject = require('terriajs-cesium/Source/Core/destroyObject');
 var JulianDate = require('terriajs-cesium/Source/Core/JulianDate');
 var loadText = require('terriajs-cesium/Source/Core/loadText');
 
-var formatPropertyValue = require('../Core/formatPropertyValue');
+var formatPropertyValue = require('../Models/formatPropertyValue');
 var DataTable = require('./DataTable');
 
 /**
