+++ lib/Models/GeoJsonCatalogItem.js
@@ -25,7 +25,7 @@ var PointGraphics = require('terriajs-cesium/Source/DataSources/PointGraphics');
 var Metadata = require('./Metadata');
 var ModelError = require('./ModelError');
 var CatalogItem = require('./CatalogItem');
-var formatPropertyValue = require('../Core/formatPropertyValue');
+var formatPropertyValue = require('../Models/formatPropertyValue');
 var inherit = require('../Core/inherit');
 var Proj4Definitions = require('../Map/Proj4Definitions');
 var proxyCatalogItemUrl = require('./proxyCatalogItemUrl');
