+++ lib/Models/ImageryLayerFeatureInfo.js
@@ -4,7 +4,7 @@
 var ImageryLayerFeatureInfo = require('terriajs-cesium/Source/Scene/ImageryLayerFeatureInfo');
 var defined = require('terriajs-cesium/Source/Core/defined');
 
-var formatPropertyValue = require('../Core/formatPropertyValue');
+var formatPropertyValue = require('../Models/formatPropertyValue');
 
 /**
  * Configures the description of this feature by creating an HTML table of properties and their values.
