+++ lib/Models/formatPropertyValue.js
